backend/auth/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pymongo==4.10.1
dnspython==2.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic[email]==2.10.0
python-multipart==0.0.17
python-dotenv==1.0.0
certifi==2024.8.30
